Player movement improvement + design considerations
 ### tl;dr: 

1. Static methods (clamp() => Vector2.clamp())
2. Don't put grid stuff in player (ECS good)
 
Later, when we have an ECS (Entity Component System) everything should be
broken up in to appropriate 'boxes'. For player movement, we'd need:

- **Entity**: Player
  - _Component_: Position
- **Entity**: Grid
  - _Component_: Area => (Size, Bounds)
-  **System**: Movement(Position, Actions)
   - _Subsystem_: BoundedMovment(Area, Position)
  
 So the logically, the properties involving the grid
 (like cell offset, grid size, etc) would not be in the player movement.


An example benefit of such an ECS would be a "jail" that limits player movement more. Instead of processing the bounds twice, we can create a method on Area to clamp to the smallest one of the given Areas. 
Then in the Movement system, we would process it over the player only once, instead of multiple times, costing over head and potential bugs because logical order.

Additionally, if we needed the same setup for a non-player Entity, we could reuse the same system, and even query the number Entities trapped by the jail!
	
For now though, I've statically typed and
separated variables here for clarity, since:
- The math makes it look like a scalar (normal number), not a vector
- It's harder to know what it's doing without the comment
- The name "border" isn't informative