Pull request for dot movement (#6)
* Player now moves on (WASD) press

* Added grid and made movement code worse

* Removed unnecessary shader

* Optimized movement code

* Trying to get key collision

* key no longer hidden by defauly

* Now handles movement in _input

* Auto stash before merge of "feature/dot_movement" and "origin/feature/dot_movement"

* Grid now generates on start and player starts at 1,1

* Added arrow keys to project settings

* Rewrite of movement block

Fixes #4

* Grid now scales

You can now change the x and y dimensions of the grid and change the size of each tile

* Player movement and size now scales with grid

* Player now stops of edge of grid

* Player movement improvement + design considerations

 ### tl;dr: 

1. Static methods (clamp() => Vector2.clamp())
2. Don't put grid stuff in player (ECS good)
 
Later, when we have an ECS (Entity Component System) everything should be
broken up in to appropriate 'boxes'. For player movement, we'd need:

- **Entity**: Player
  - _Component_: Position
- **Entity**: Grid
  - _Component_: Area => (Size, Bounds)
-  **System**: Movement(Position, Actions)
   - _Subsystem_: BoundedMovment(Area, Position)
  
 So the logically, the properties involving the grid
 (like cell offset, grid size, etc) would not be in the player movement.


An example benefit of such an ECS would be a "jail" that limits player movement more. Instead of processing the bounds twice, we can create a method on Area to clamp to the smallest one of the given Areas. 
Then in the Movement system, we would process it over the player only once, instead of multiple times, costing over head and potential bugs because logical order.

Additionally, if we needed the same setup for a non-player Entity, we could reuse the same system, and even query the number Entities trapped by the jail!
	
For now though, I've statically typed and
separated variables here for clarity, since:
- The math makes it look like a scalar (normal number), not a vector
- It's harder to know what it's doing without the comment
- The name "border" isn't informative

* Improved directory structure

- Added assets structure
- Added ".ignore" folder (see .gitignore)

* Player now moves on (WASD) press

* Added grid and made movement code worse

* Removed unnecessary shader

* Optimized movement code

* Trying to get key collision

* key no longer hidden by defauly

* Now handles movement in _input

* Auto stash before merge of "feature/dot_movement" and "origin/feature/dot_movement"

* Grid now generates on start and player starts at 1,1

* Rewrite of movement block

Fixes #4

* Grid now scales

You can now change the x and y dimensions of the grid and change the size of each tile

* Player movement and size now scales with grid

* Player now stops of edge of grid

* Fixed border shit BOUNDRY

* Player now moves on (WASD) press

* Added grid and made movement code worse

* Removed unnecessary shader

* Optimized movement code

* Trying to get key collision

* key no longer hidden by defauly

* Now handles movement in _input

* Grid now generates on start and player starts at 1,1

* Added arrow keys to project settings

* Rewrite of movement block

Fixes #4

* Grid now scales

You can now change the x and y dimensions of the grid and change the size of each tile

* Player movement and size now scales with grid

* Player now stops of edge of grid

* Player movement improvement + design considerations

 ### tl;dr: 

1. Static methods (clamp() => Vector2.clamp())
2. Don't put grid stuff in player (ECS good)
 
Later, when we have an ECS (Entity Component System) everything should be
broken up in to appropriate 'boxes'. For player movement, we'd need:

- **Entity**: Player
  - _Component_: Position
- **Entity**: Grid
  - _Component_: Area => (Size, Bounds)
-  **System**: Movement(Position, Actions)
   - _Subsystem_: BoundedMovment(Area, Position)
  
 So the logically, the properties involving the grid
 (like cell offset, grid size, etc) would not be in the player movement.


An example benefit of such an ECS would be a "jail" that limits player movement more. Instead of processing the bounds twice, we can create a method on Area to clamp to the smallest one of the given Areas. 
Then in the Movement system, we would process it over the player only once, instead of multiple times, costing over head and potential bugs because logical order.

Additionally, if we needed the same setup for a non-player Entity, we could reuse the same system, and even query the number Entities trapped by the jail!
	
For now though, I've statically typed and
separated variables here for clarity, since:
- The math makes it look like a scalar (normal number), not a vector
- It's harder to know what it's doing without the comment
- The name "border" isn't informative

* Player now moves on (WASD) press

* Added grid and made movement code worse

* Removed unnecessary shader

* Optimized movement code

* Trying to get key collision

* key no longer hidden by defauly

* Now handles movement in _input

* Auto stash before merge of "feature/dot_movement" and "origin/feature/dot_movement"

* Grid now generates on start and player starts at 1,1

* Rewrite of movement block

Fixes #4

* Grid now scales

You can now change the x and y dimensions of the grid and change the size of each tile

* Player movement and size now scales with grid

* Player now stops of edge of grid

* Fixed border shit BOUNDRY

---------

Co-authored-by: wetda3 <124001666+sdemayo@users.noreply.github.com>
Co-authored-by: Tahv Demayo <tahvsprojects@gmail.com>
Co-authored-by: werydude <werydude@users.noreply.github.com>